refactor(list): icons moved from src
ISSUES CLOSED: #81
