+++ list/src/Item.tsx
@@ -2,7 +2,7 @@ import React, { useState }                from 'react'
 
 import { Button }                         from './button'
 import { Condition }                      from './condition'
-import { HeartIcon, StarIcon, TrashIcon } from './icons/src'
+import { HeartIcon, StarIcon, TrashIcon } from './icons'
 import { Box, Column, Layout, Row }       from './layout'
 import { Text }                           from './text'
