+++ list/src/icons/StarIcon.tsx
