+++ list/src/icons/TrashIcon.tsx
